paginacion a las tablas

+++ src/main/java/com/proyecto/PoryectoBuzu/controllers/ProductosController.java
@@ -2,8 +2,11 @@ package com.proyecto.PoryectoBuzu.controllers;
 
 import com.proyecto.PoryectoBuzu.dao.ProductosDao;
 
+import com.proyecto.PoryectoBuzu.models.Errores;
 import com.proyecto.PoryectoBuzu.models.Productos;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -29,20 +32,39 @@ public class ProductosController {
 
 
     @RequestMapping(value = "api/productos", method = RequestMethod.POST)
-    public void registrarEmpleado(@RequestParam("productoImagen") MultipartFile imagen,
-                                  @RequestParam("sku") Long sku,
-                                  @RequestParam("nombre") String nombre,
-                                  @RequestParam("descripcion") String descripcion,
-                                  @RequestParam("cantidad") int cantidad,
-                                  @RequestParam("categoria") String categoria,
-                                  @RequestParam("precioCompra") double precioCompra,
-                                  @RequestParam("descuento") double descuento,
-                                  @RequestParam("proveedor") String proveedor,
-                                  @RequestParam("resumen_prod") String resumen){
+    public Errores registrarEmpleado(@RequestParam("productoImagen") MultipartFile imagen,
+                                            @RequestParam("sku") String sku,
+                                            @RequestParam("nombre") String nombre,
+                                            @RequestParam("descripcion") String descripcion,
+                                            @RequestParam("cantidad") int cantidad,
+                                            @RequestParam("categoria") String categoria,
+                                            @RequestParam("precioCompra") double precioCompra,
+                                            @RequestParam("descuento") double descuento,
+                                            @RequestParam("proveedor") String proveedor,
+                                            @RequestParam("resumen_prod") String resumen){
+
+        Errores errores = new Errores();
 
 
-        if(!imagen.isEmpty()){
+        if (sku == null) {
+            errores.setCodigoSKU("El código SKU es incorrecto");
+        }
+        if ("FAIL".equals(productoDao.verificarCodigo(sku.toUpperCase()))) {
+            errores.setCodigoSKU("El código SKU ya se encuentra registrado");
+        }
 
+        errores.setImagen(imagen.isEmpty() ? "Debe seleccionar una imagen" : null);
+        errores.setNombre(nombre.isEmpty() ? "Debe ingresar un nombre válido" : null);
+        errores.setDescripcion(descripcion.isEmpty() ? "Debe ingresar una descripción válida" : null);
+        errores.setCategoria(categoria.isEmpty() ? "Debe ingresar una categoría válida" : null);
+        errores.setResumen(resumen.isEmpty() ? "Debe ingresar un resumen válido" : null);
+        errores.setProveedor(proveedor.isEmpty() ? "Debe ingresar un proveedor válido" : null);
+        errores.setCantidad(cantidad <= 0 ? "La cantidad debe ser mayor que 0" : null);
+        errores.setDescuento(descuento <= 0 ? "El descuento debe ser mayor que 0" : (descuento > 90 ? "El descuento no puede superar el 90%" : null));
+
+        if (!productoDao.verificarVacio(errores)) {
+            return errores;
+        } else {
 
             String rutaAbsoluta = "images//Productos//";
             String nuevoNombreImagen = sku.toString().replaceAll(" ","") + ".jpg";
@@ -55,15 +77,12 @@ public class ProductosController {
                 throw new RuntimeException(e);
             }
 
-
-
             double precioVenta = precioCompra -  (precioCompra*descuento/100);
-
 
             // Crear un objeto Empleado con los demás datos del formulario
             Productos producto = new Productos();
             producto.setNom_prod(nombre);
-            producto.setSku_prod(sku);
+            producto.setSku_prod(sku.toUpperCase());
             producto.setDescrp_prod(descripcion);
             producto.setCantidad_prod(cantidad);
             producto.setCateg_prod(categoria);
@@ -75,14 +94,16 @@ public class ProductosController {
             producto.setResumen_product(resumen);
 
             productoDao.registrarProducto(producto);
+            errores.setExito("OK");
 
+            return errores;
         }
 
 
     }
 
     @RequestMapping(value = "api/productos/{id}", method = RequestMethod.PUT)
-    public void editarEmpleado(   @PathVariable("id") Long id,
+    public Errores editarEmpleado(   @PathVariable("id") Long id,
                                  @RequestParam("productoImagen") MultipartFile imagen,
                                   @RequestParam("nombre") String nombre,
                                   @RequestParam("descripcion") String descripcion,
@@ -90,48 +111,63 @@ public class ProductosController {
                                   @RequestParam("precioCompra") double precioCompra,
                                   @RequestParam("descuento") double descuento,
                                   @RequestParam("proveedor") String proveedor,
-                                  @RequestParam("resumen_prod") String resumen){
+                                  @RequestParam("resumen_prod") String resumen) {
 
-        Productos productoExistente = productoDao.obtenerDatosProducto(id);
 
-        String nombreImagen = productoExistente.getImg_prod();
+        Errores errores = new Errores();
 
-        String rutaExistente = "images//Productos//";
-        String rutaImagen = rutaExistente + "//" + nombreImagen;
-        File archivoImagen = new File(rutaImagen);
-        archivoImagen.delete();
+        errores.setImagen(imagen.isEmpty() ? "Debe seleccionar una imagen" : null);
+        errores.setNombre(nombre.isEmpty() ? "Debe ingresar un nombre válido" : null);
+        errores.setDescripcion(descripcion.isEmpty() ? "Debe ingresar una descripción válida" : null);
+        errores.setCategoria(categoria.isEmpty() ? "Debe ingresar una categoría válida" : null);
+        errores.setResumen(resumen.isEmpty() ? "Debe ingresar un resumen válido" : null);
+        errores.setProveedor(proveedor.isEmpty() ? "Debe ingresar un proveedor válido" : null);
+        errores.setDescuento(descuento <= 0 ? "El descuento debe ser mayor que 0" : (descuento > 90 ? "El descuento no puede superar el 90%" : null));
 
-        if(imagen != null && !imagen.isEmpty()) {
-            // Procesar la nueva imagen solo si se proporciona una imagen válida
 
-            String nuevoNombreImagen = productoExistente.getSku_prod()+".jpg";
+        if (!productoDao.verificarVacio(errores)) {
+            return errores;
+        } else {
 
-            try {
-                byte[] bytesImg = imagen.getBytes();
-                Path rutacompleta = Paths.get(rutaExistente + "//" + nuevoNombreImagen);
-                Files.write(rutacompleta, bytesImg);
-                productoExistente.setImg_prod(nuevoNombreImagen);
-            } catch (Exception e) {
-                throw new RuntimeException(e);
-            }
+            Productos productoExistente = productoDao.obtenerDatosProducto(id);
+
+            String nombreImagen = productoExistente.getImg_prod();
+
+            String rutaExistente = "images//Productos//";
+            String rutaImagen = rutaExistente + "//" + nombreImagen;
+            File archivoImagen = new File(rutaImagen);
+            archivoImagen.delete();
+
+                String nuevoNombreImagen = productoExistente.getSku_prod() + ".jpg";
+
+                try {
+                    byte[] bytesImg = imagen.getBytes();
+                    Path rutacompleta = Paths.get(rutaExistente + "//" + nuevoNombreImagen);
+                    Files.write(rutacompleta, bytesImg);
+                    productoExistente.setImg_prod(nuevoNombreImagen);
+                } catch (Exception e) {
+                    throw new RuntimeException(e);
+                }
+
+
+                double precioVenta = precioCompra - (precioCompra * descuento / 100);
+
+                productoExistente.setNom_prod(nombre);
+                productoExistente.setDescrp_prod(descripcion);
+                productoExistente.setCateg_prod(categoria);
+                productoExistente.setCompra(precioCompra);
+                productoExistente.setVenta(precioVenta);
+                productoExistente.setDescuento(descuento);
+                productoExistente.setProveedor(proveedor);
+                productoExistente.setResumen_product(resumen);
+
+                productoDao.editarProducto(productoExistente);
+                errores.setExito("OK");
+
+                return errores;
+
         }
-
-
-        double precioVenta = precioCompra -  (precioCompra*descuento/100);
-
-
-        productoExistente.setNom_prod(nombre);
-        productoExistente.setDescrp_prod(descripcion);
-        productoExistente.setCateg_prod(categoria);
-        productoExistente.setCompra(precioCompra);
-        productoExistente.setVenta(precioVenta);
-        productoExistente.setDescuento(descuento);
-        productoExistente.setProveedor(proveedor);
-        productoExistente.setResumen_product(resumen);
-
-        productoDao.editarProducto(productoExistente);
-
-        }
+    }
 
 
         @RequestMapping(value = "api/productos/{idProd}", method = RequestMethod.GET)
